Add mutate_ConnectNeurons to base Mutagen class and remove redundant type conversions
- Added ConnectNeurons clause to Mutagen.mutate() method
- Added mutate_ConnectNeurons() method to base Mutagen class
- Refactored ModifyConnectionMutagen to override mutate_ConnectNeurons()
- Added type hints to mutate_ConnectNeurons() method
- Removed redundant .astype(int) calls in DuplicateGeneMutagen and ResizeNeuronLayerMutagen
- All 20 tests passing for affected mutagens

+++ src/roxene/mutagens/duplicate_gene_mutagen.py
@@ -43,7 +43,7 @@ class DuplicateGeneMutagen(InsertGeneToCompositeMutagen):
 
         # Duplicate a random gene
         if len(new_genes) > 0:
-            index_to_duplicate = get_rng().integers(0, len(new_genes)).astype(int)
+            index_to_duplicate = get_rng().integers(0, len(new_genes))
             # Insert the duplicate right after the original
             new_genes.insert(index_to_duplicate + 1, new_genes[index_to_duplicate])
